Add CheckTractorServiceAlerts command and related job for service alerts
- Introduced a new console command `CheckTractorServiceAlerts` to check periodic service thresholds for tractors and send alerts.
- Created `CheckTractorServiceAlertsJob` to handle the logic for checking service alerts and notifying users.
- Added `AppReleaseController`, `StoreAppReleaseRequest`, and `AppReleaseResource` for managing application releases, including validation and resource formatting.
- Updated `LoadEstimationController` and `MaintenanceReportController` to include new fields and validation rules for better data handling.
- Enhanced `NutrientDiagnosisRequest` and related policies to support new update and approval workflows.
- Implemented `MaintenanceDistanceMonitorService` to monitor maintenance distance thresholds and notify farm admins accordingly.
- Refactored various models and resources to accommodate new fields and relationships, improving overall data integrity and API responses.

+++ app/Http/Controllers/Api/V1/Farm/ColdRequirementController.php
@@ -62,12 +62,12 @@ class ColdRequirementController extends Controller
      * Calculate the cold requirement for the farm.
      *
      * @param array $data
-     * @param int $minTemp
-     * @param int $maxTemp
+     * @param float $minTemp
+     * @param float $maxTemp
      * @param string $method
      * @return int
      */
-    private function calculateColdRequirement(array $data, int $minTemp, int $maxTemp, string $method): int
+    private function calculateColdRequirement(array $data, float $minTemp, float $maxTemp, string $method): int
     {
         return $method === 'method1'
             ? $this->calculateColdRequirementMethod1($data, $minTemp, $maxTemp)
@@ -78,11 +78,11 @@ class ColdRequirementController extends Controller
      * Calculate the cold requirement for the farm using method 1.
      *
      * @param array $data
-     * @param int $minTemp
-     * @param int $maxTemp
+     * @param float $minTemp
+     * @param float $maxTemp
      * @return int
      */
-    private function calculateColdRequirementMethod1(array $data, int $minTemp, int $maxTemp): int
+    private function calculateColdRequirementMethod1(array $data, float $minTemp, float $maxTemp): int
     {
         return collect($data['forecast']['forecastday'])->sum(function ($day) use ($minTemp, $maxTemp) {
             return collect($day['hour'])->filter(function ($hour) use ($minTemp, $maxTemp) {
